refactor schema definition to support multiple inheritance
Refactored the schema def and query/mutiation classses to support multiple inheritance. This allows individual “subservices” of the larger labbook service to be built, tested, and potentially deployed independently. Right now, we can put all the queries and mutations into a single schema so it is served from a single endpoint.